app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
requests==2.32.3
sickle==0.7.0
lxml==5.3.0
beautifulsoup4==4.12.3
pypdf==5.1.0
chromadb==0.5.5
tqdm==4.66.5
langchain==0.2.12
langchain-community==0.2.11
langchain-ollama==0.1.3
langchain-chroma
